BREAKING change app to use int instead of Double for prices. feedback system needed

+++ app/src/main/java/com/example/salestracker/data/model/Product.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 data class Product(  // Represents a product that can be sold
     @PrimaryKey val id: String = "",
     val name: String = "",
-    val defaultPrice: Double = 0.0,
+    val defaultPrice: Int = 0,
     val inStock: Boolean = false, //just added this
     val notes: String? = null,
     val orderIndex: Int = 0
