+++ app/src/main/java/com/example/salestracker/data/model/SaleItem.kt
@@ -4,7 +4,6 @@ import androidx.room.Entity
 import androidx.room.ForeignKey
 import androidx.room.Index
 import androidx.room.PrimaryKey
-import com.example.salestracker.SaleEvent
 
 @Entity(tableName = "sale_item",
     foreignKeys = [
@@ -32,6 +31,6 @@ data class SaleItem(   // Represents one item in a sale (where one sale can have
     @PrimaryKey val id: String,
     val saleId: String,
     val productId: String,
-    val salePrice: Double, //represents the price of the sale item. could be quantity 5 were sold, making 5 x Product.price
+    val salePrice: Int, //represents the price of the sale item. could be quantity 5 were sold, making 5 x Product.price
     val quantity: Int
 )
